claimDateFrom parameter not working, performance
The claimDateFrom parameter was not properly handled (converted to date but the query still used the text version). The query was also implemented on a "in" from the Claims table which would have been terribly inefficient.

I do not like the date format either: DD-MM-YYYY instead of YYYY-MM-DD but I didn't want to break software relying on this.