Switch FHIR security to the OpenIMIS model ⚠️
So far, the FHIR API was allowed or denied based on Django permissions per module, model and method. This didn't filter out the data that was out of the user's range.
The authorisations are now taken from the same source as the GraphQL API. It means that it might not be a 100% match.

There might still be a need to adjust some permission mappings or some restrictions.

The automated tests were not yet updated for this change, it was tested manually.