OMT-204 remove eval()
When we removed the unused imports to please LGTM, the FHIR module stopped working. This was because it took the serializer class from the configuration with an eval() call. This meant that it relied on having those unused imports and, although the configuration is not accessible to everyone, it remains a dangerous code execution from a string.
getattr() is an excellent replacement as it provides the same service, without any risk and without the need for the unused imports